requirements.txt
Django==1.8.4
django-debug-toolbar==1.3.2
django-localflavor==1.1
django-mongodbforms==0.3.1
mongodbforms==0.3
mongoengine==0.10.0
pymongo==3.0.3
python-dateutil==2.4.2
six==1.9.0
